feat: implemented getTop10Quotes & getWorst10Quotes methods

+++ src/main/java/com/example/test_kameleoon/models/User.java
@@ -15,11 +15,9 @@ import java.util.Date;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String name;
     private String email;
     private String password;
     private Date dateOfCreation;
-
-
 }
